# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Employee Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
...
  /branches:
    get:
      operationId: listBranches
      tags: [branches]
      summary: List branches
      responses:
        "200":
          description: List of branches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Branch"
    post:
      operationId: createBranch
      tags: [branches]
      summary: Create branch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
  /branches/{id}:
    patch:
      operationId: updateBranch
      tags: [branches]
      summary: Update branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BranchUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branch"
    delete:
      operationId: deleteBranch
      tags: [branches]
      summary: Delete branch
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      summary: List attendance records
      parameters:
        - name: employeeId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
        - name: endDate
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Attendance"
    post:
      operationId: createAttendance
      tags: [attendance]
      summary: Create manual attendance record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /attendance/summary:
    get:
      operationId: getAttendanceSummary
      tags: [attendance]
      summary: Get attendance summary
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
        - name: employeeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Attendance summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AttendanceSummary"
...
  /attendance/web-punch:
    post:
      operationId: webPunch
      tags: [attendance]
      summary: Employee self web punch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebPunchInput"
      responses:
        "201":
          description: Punched
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
  /attendance/zkteco:
    post:
      operationId: zktecoWebhook
      tags: [attendance]
      summary: ZKTeco ADMS webhook endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZktecoPayload"
      responses:
        "200":
          description: Received
  /attendance/{id}:
    patch:
      operationId: updateAttendance
      tags: [attendance]
      summary: Update/correct attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttendanceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attendance"
    delete:
      operationId: deleteAttendance
      tags: [attendance]
      summary: Delete attendance record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        token:
          type: string
    UserProfile:
      type: object
      required: [id, email, role, name]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        employeeId:
          type: ["integer", "null"]
...
    Branch:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    BranchInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        phone:
          type: string
    BranchUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        city:
          type: string
        phone:
          type: string
        isActive:
          type: boolean
...
    Attendance:
      type: object
      required: [id, employeeId, date, createdAt]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        employeeCode:
          type: ["string", "null"]
        date:
          type: string
        checkIn:
          type: ["string", "null"]
        checkOut:
          type: ["string", "null"]
        workingHours:
          type: ["number", "null"]
        status:
          type: string
        isLate:
          type: boolean
        isEarlyOut:
          type: boolean
        source:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    AttendanceInput:
      type: object
      required: [employeeId, date]
      properties:
        employeeId:
          type: integer
        date:
          type: string
        checkIn:
          type: string
        checkOut:
          type: string
        status:
          type: string
        notes:
          type: string
    AttendanceUpdate:
      type: object
      properties:
        checkIn:
          type: string
        checkOut:
          type: string
        status:
          type: string
        notes:
          type: string
    AttendanceSummary:
      type: object
      required: [totalDays, presentDays, absentDays, lateDays, leaveDays, averageWorkingHours]
      properties:
        totalDays:
          type: integer
        presentDays:
          type: integer
        absentDays:
          type: integer
        lateDays:
          type: integer
        leaveDays:
          type: integer
        averageWorkingHours:
          type: number
        attendanceRate:
          type: number
    WebPunchInput:
      type: object
      required: [punchType]
      properties:
        punchType:
          type: string
    ZktecoPayload:
      type: object
      required: [sn, table]
      properties:
        sn:
          type: string
        table:
          type: string
        Stamp:
          type: ["string", "null"]
        OpStamp:
          type: ["string", "null"]
        templateDataList:
          type: ["string", "null"]
...