# lib/api-zod/src/generated/types/branch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Employee Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Branch {
  id: number;
  name: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  phone?: string | null;
  isActive?: boolean;
  createdAt: string;
}