# lib/api-zod/src/generated/types/branchUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Employee Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface BranchUpdate {
  name?: string;
  address?: string;
  city?: string;
  phone?: string;
  isActive?: boolean;
}